Add style to the action section with CSS
Since we used tags like the button and left right div classes again in this section, we had to add some
specificity in the hero and the action section in the CSS for these selectors, since now they require
different CSS configurations. The button in action is just like the one in hero but
with border. The colors and fonts in this section all follow styles already previously
used in the page.